% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_consolidate.R
\name{detect_semester_from_canvas}
\alias{detect_semester_from_canvas}
\title{Detect semester from Canvas data}
\usage{
detect_semester_from_canvas(canvas_data)
}
\arguments{
\item{canvas_data}{data.frame from import_canvas_grades with semester attribute}
}
\value{
Character semester code (e.g., "S1", "S2", "S2C")
}
\description{
Detect semester from Canvas data
}
\keyword{internal}
