% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_consolidate.R
\name{detect_unit_from_canvas}
\alias{detect_unit_from_canvas}
\title{Detect unit code from Canvas data}
\usage{
detect_unit_from_canvas(canvas_data)
}
\arguments{
\item{canvas_data}{data.frame from import_canvas_grades}
}
\value{
Character unit code (e.g., "BIOL2022")
}
\description{
Detect unit code from Canvas data
}
