% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_consolidate.R
\name{detect_year_from_canvas}
\alias{detect_year_from_canvas}
\title{Detect year from Canvas data}
\usage{
detect_year_from_canvas(canvas_data)
}
\arguments{
\item{canvas_data}{data.frame from import_canvas_grades}
}
\value{
Character year code (e.g., "2025")
}
\description{
Detect year from Canvas data
}
