% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\docType{data}
\name{NIGHTMARE_CONFIG}
\alias{NIGHTMARE_CONFIG}
\title{Application configuration}
\format{
An object of class \code{list} of length 3.
}
\usage{
NIGHTMARE_CONFIG
}
\description{
Application configuration
}
\keyword{datasets}
