% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_folder_loader.R
\name{read_last_unit}
\alias{read_last_unit}
\title{Read the last loaded unit from disk}
\usage{
read_last_unit(data_dir)
}
\arguments{
\item{data_dir}{Path to the root data directory}
}
\value{
Character string of last unit name, or NULL
}
\description{
Read the last loaded unit from disk
}
\keyword{internal}
