% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_exam.R
\name{read_tabular}
\alias{read_tabular}
\title{Read tabular file (CSV or Excel)}
\usage{
read_tabular(path, ...)
}
\arguments{
\item{path}{Path to file}

\item{...}{Additional arguments passed to reader}
}
\value{
data.frame
}
\description{
Read tabular file (CSV or Excel)
}
\keyword{internal}
