% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_nightmare}
\alias{run_nightmare}
\title{Run the NIGHTMARE Shiny app}
\usage{
run_nightmare(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny]{runApp}}}
}
\description{
Run the NIGHTMARE Shiny app
}
