% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_folder_loader.R
\name{save_last_unit}
\alias{save_last_unit}
\title{Save the last loaded unit to disk}
\usage{
save_last_unit(data_dir, unit)
}
\arguments{
\item{data_dir}{Path to the root data directory}

\item{unit}{Character string of unit name}
}
\description{
Save the last loaded unit to disk
}
